+++ www/worker-compositor.js
@@ -7,7 +7,7 @@ let size_y;
 function compose(workerResultBuffer) {
     console.log("compositor recieved data to compose");
     
-    const workerResult = new Uint32Array(workerResultBuffer);
+    const workerResult = new Uint16Array(workerResultBuffer);
 
     let min = Number.MAX_SAFE_INTEGER;
     let max = 0;
@@ -79,7 +79,7 @@ function init(s_x, s_y) {
     size_x = s_x;
     size_y = s_y;
 	console.log("initializing compositor with values: "+size_x+"x"+size_y);
-	cumulatedResult = new Uint32Array(size_x*size_y);
+	cumulatedResult = new Uint16Array(size_x*size_y);
     cumulatedResult.fill(0);
 }
